docs: add filter jsdoc
