+++ packages/sheets-filter/src/facade/f-event.ts
@@ -14,44 +14,63 @@
  * limitations under the License.
  */
 
-import type { IEventBase } from '@univerjs/core';
+import type { ICommandInfo, IEventBase, Injector } from '@univerjs/core';
+import type { ISheetCommandSharedParams } from '@univerjs/sheets';
 import type { ISetSheetsFilterCriteriaCommandParams } from '@univerjs/sheets-filter';
 import type { FWorkbook, FWorksheet } from '@univerjs/sheets/facade';
-import { FEventName } from '@univerjs/core';
+import { FEventName, FUniver, ICommandService } from '@univerjs/core';
+
+import { ClearSheetsFilterCriteriaCommand, SetSheetsFilterCriteriaCommand } from '@univerjs/sheets-filter';
 import { FSheetEventName } from '@univerjs/sheets/facade';
 
 // TODO: add js comments for the following API
 
 export interface IFSheetFilterEventMixin {
     /**
-     * Type of the event's parameters is {@link ISheetRangeFilteredParams}.
+     * This event will be emitted when the filter criteria on a column is changed.
+     * Type of the event is {@link ISheetRangeFilteredParams}.
      *
-     * ```typescript
      * @example
+     * ```typescript
+     * const callbackDisposable = univerAPI.addEvent(univerAPI.Event.SheetRangeFiltered, (event) => {
+     *   const { workbook, worksheet, col, criteria } = params;
+     * });
      * ```
      */
     SheetRangeFiltered: 'SheetRangeFiltered';
     /**
-     * Type of the event's parameters is {@link ISheetRangeFilteredParams}.
+     * This event will be emitted beftore the filter criteria on a column is changed.
+     * Type of the event is {@link ISheetRangeFilteredParams}.
      *
-     * ```typescript
      * @example
+     * ```typescript
+     * const callbackDisposable = univerAPI.addEvent(univerAPI.Event.SheetBeforeRangeFilter, (event) => {
+     *   const { workbook, worksheet, col, criteria } = params;
+     * });
      * ```
      */
     SheetBeforeRangeFilter: 'SheetBeforeRangeFilter';
     /**
-     * Type of the event's parameters is {@link ISheetRangeFilterClearedEventParams}.
+     * This event will be emitted when the filter on a worksheet is cleared.
+     * Type of the event is {@link ISheetRangeFilterClearedEventParams}.
      *
-     * ```typescript
      * @example
+     * ```typescript
+     * const callbackDisposable = univerAPI.addEvent(univerAPI.Event.SheetRangeFilterCleared, (event) => {
+     *   const { workbook, worksheet } = params;
+     * });
      * ```
      */
     SheetRangeFilterCleared: 'SheetRangeFilterCleared';
     /**
-     * Type of the event's parameters is {@link ISheetRangeFilterClearedEventParams}.
+     * This event will be emitted after the filter on a worksheet is cleared.
+     * Type of the event is {@link ISheetRangeFilterClearedEventParams}.
      *
-     * ```typescript
      * @example
+     * ```typescript
+     * const callbackDisposable = univerAPI.addEvent(univerAPI.Event.SheetBeforeRangeFilterClear, (event) => {
+     *   const { workbook, worksheet } = params;
+     * });
      * ```
      */
     SheetBeforeRangeFilterClear: 'SheetBeforeRangeFilterClear';
@@ -88,3 +107,92 @@ declare module '@univerjs/core' {
     interface IEventParamConfig extends ISheetRangeFilterEventParamConfig { }
 }
 
+class FUniverSheetsFilterEventMixin extends FUniver {
+    override _initialize(injector: Injector): void {
+        const commandService = injector.get(ICommandService);
+
+        this.disposeWithMe(commandService.beforeCommandExecuted((commandInfo) => {
+            switch (commandInfo.id) {
+                case SetSheetsFilterCriteriaCommand.id:
+                    this._beforeRangeFilter(commandInfo as Readonly<ICommandInfo<ISetSheetsFilterCriteriaCommandParams>>);
+                    break;
+                case ClearSheetsFilterCriteriaCommand.id:
+                    this._beforeRangeFilterClear(commandInfo as Readonly<ICommandInfo<ISheetCommandSharedParams>>);
+                    break;
+            }
+        }));
+
+        this.disposeWithMe(commandService.onCommandExecuted((commandInfo) => {
+            switch (commandInfo.id) {
+                case SetSheetsFilterCriteriaCommand.id:
+                    this._onRangeFiltered(commandInfo as Readonly<ICommandInfo<ISetSheetsFilterCriteriaCommandParams>>);
+                    break;
+                case ClearSheetsFilterCriteriaCommand.id:
+                    this._onRangeFilterCleared(commandInfo as Readonly<ICommandInfo<ISheetCommandSharedParams>>);
+                    break;
+            }
+        }));
+    }
+
+    private _beforeRangeFilter(commandInfo: Readonly<ICommandInfo<ISetSheetsFilterCriteriaCommandParams>>): void {
+        const params = commandInfo.params!;
+        const fWorkbook = this.getUniverSheet(params.unitId)!;
+        const eventParams: ISheetRangeFilteredParams = {
+            workbook: fWorkbook,
+            worksheet: fWorkbook.getSheetBySheetId(params.subUnitId)!,
+            col: params.col,
+            criteria: params.criteria,
+        };
+
+        this.fireEvent(this.Event.SheetBeforeRangeFilter, eventParams);
+        if (eventParams.cancel) {
+            throw new Error('SetSheetsFilterCriteriaCommand canceled.');
+        }
+    }
+
+    private _onRangeFiltered(commandInfo: Readonly<ICommandInfo<ISetSheetsFilterCriteriaCommandParams>>): void {
+        const params = commandInfo.params!;
+        const fWorkbook = this.getUniverSheet(params.unitId)!;
+        const eventParams: ISheetRangeFilteredParams = {
+            workbook: fWorkbook,
+            worksheet: fWorkbook.getSheetBySheetId(params.subUnitId)!,
+            col: params.col,
+            criteria: params.criteria,
+        };
+
+        this.fireEvent(this.Event.SheetRangeFiltered, eventParams);
+        if (eventParams.cancel) {
+            throw new Error('SetSheetsFilterCriteriaCommand canceled.');
+        }
+    }
+
+    private _beforeRangeFilterClear(commandInfo: Readonly<ICommandInfo<ISheetCommandSharedParams>>): void {
+        const params = commandInfo.params!;
+        const fWorkbook = this.getUniverSheet(params.unitId)!;
+        const eventParams: ISheetRangeFilterClearedEventParams = {
+            workbook: fWorkbook,
+            worksheet: fWorkbook.getSheetBySheetId(params.subUnitId)!,
+        };
+
+        this.fireEvent(this.Event.SheetBeforeRangeFilterClear, eventParams);
+        if (eventParams.cancel) {
+            throw new Error('SetSheetsFilterCriteriaCommand canceled.');
+        }
+    }
+
+    private _onRangeFilterCleared(commandInfo: Readonly<ICommandInfo<ISheetCommandSharedParams>>): void {
+        const params = commandInfo.params!;
+        const fWorkbook = this.getUniverSheet(params.unitId)!;
+        const eventParams: ISheetRangeFilterClearedEventParams = {
+            workbook: fWorkbook,
+            worksheet: fWorkbook.getSheetBySheetId(params.subUnitId)!,
+        };
+
+        this.fireEvent(this.Event.SheetRangeFilterCleared, eventParams);
+        if (eventParams.cancel) {
+            throw new Error('SetSheetsFilterCriteriaCommand canceled.');
+        }
+    }
+}
+
+FUniver.extend(FUniverSheetsFilterEventMixin);
